% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets_used.R
\name{datasets_groups}
\alias{datasets_groups}
\title{Datasets List used in RnBeads Analysis}
\usage{
datasets_groups(rd)
}
\arguments{
\item{rd}{A path to the rnbeads results repository}
}
\value{
list of directories
}
\description{
Takes in a results directory and returns a list of grouped directories that used the same set of datasets
}

