% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_tables.R
\name{datasets_list}
\alias{datasets_list}
\title{Different Datasets used in RnBeads Analysis}
\usage{
datasets_list(rd)
}
\arguments{
\item{rd}{A path to the rnbeads results repository}
}
\value{
list of datasets
}
\description{
Takes in a results directory and returns a list of different datasets used
}

