% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_datasets.R
\name{func_common_datasets}
\alias{func_common_datasets}
\title{Datasets List used in RnBeads Analysis}
\usage{
func_common_datasets(rd)
}
\arguments{
\item{rd}{A path to the rnbeads results repository}
}
\value{
list of directories
}
\description{
Takes in a results directory and returns a list of grouped directories that used the same set of datasets
}

