% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{squareri}
\alias{squareri}
\title{Square a number}
\usage{
squareri(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
Squared value of x
}
\description{
Takes in any numeric value and square it
}

